refactor core functionality and models into separate folder to be able to pull it up later
