+++ pkg/core/models/models.go
@@ -1,15 +1,24 @@
 package models
 
-import "digger/pkg/configuration"
-
 type ProjectCommand struct {
 	ProjectName      string
 	ProjectDir       string
 	ProjectWorkspace string
 	Terragrunt       bool
 	Commands         []string
-	ApplyStage       *configuration.Stage
-	PlanStage        *configuration.Stage
+	ApplyStage       *Stage
+	PlanStage        *Stage
 	StateEnvVars     map[string]string
 	CommandEnvVars   map[string]string
+}
+
+type Step struct {
+	Action    string
+	Value     string
+	ExtraArgs []string
+	Shell     string
+}
+
+type Stage struct {
+	Steps []Step
 }
