+++ pkg/core/reporting/reporting.go
@@ -1,0 +1,5 @@
+package reporting
+
+type Reporter interface {
+	Report(report string) error
+}
