+++ pkg/integration/integration_test.go
@@ -4,12 +4,12 @@ import (
 	"context"
 	"digger/pkg/aws"
 	"digger/pkg/configuration"
+	"digger/pkg/core/terraform"
 	"digger/pkg/digger"
 	dg_github "digger/pkg/github"
 	"digger/pkg/locking"
 	"digger/pkg/reporting"
 	"digger/pkg/storage"
-	"digger/pkg/terraform"
 	"digger/pkg/utils"
 	"log"
 	"math/rand"
