+++ pkg/storage/plan_storage.go
@@ -15,13 +15,6 @@ import (
 	"github.com/google/go-github/v51/github"
 )
 
-type PlanStorage interface {
-	StorePlan(localPlanFilePath string, storedPlanFilePath string) error
-	RetrievePlan(localPlanFilePath string, storedPlanFilePath string) (*string, error)
-	DeleteStoredPlan(storedPlanFilePath string) error
-	PlanExists(storedPlanFilePath string) (bool, error)
-}
-
 type PlanStorageGcp struct {
 	Client  *storage.Client
 	Bucket  *storage.BucketHandle
